+++ src/plugins/core/metaData/refMetaDataHandler.ts
@@ -1,6 +1,7 @@
 import { log } from '../../../io';
 import * as models from '../../../models';
 import * as utils from '../../../utils';
+import { addSkippedTestResult } from '../../../utils';
 
 export function refMetaDataHandler(type: string, name: string | undefined, context: models.ParserContext): boolean {
   if (['ref', 'forceRef'].indexOf(type) >= 0 && name) {
@@ -36,6 +37,11 @@ class RefMetaAction {
     utils.report(context, `load reference ${name}`);
     const reference = utils.findHttpRegionInContext(name, context);
     if (reference) {
+      if (this.isReferenceErroredOrSkipped(reference)) {
+        addSkippedTestResult(context.httpRegion);
+        return false;
+      }
+
       const envKey = utils.toEnvironmentKey(context.activeEnvironment);
       utils.setVariableInContext(reference.variablesPerEnv[envKey], context);
       log.trace('import variables', reference.variablesPerEnv[envKey]);
@@ -54,5 +60,11 @@ class RefMetaAction {
       throw new Error(`ref ${name} not found`);
     }
     return result;
+  }
+
+  private isReferenceErroredOrSkipped(ref: models.HttpRegion) {
+    return !!ref.testResults?.some(
+      t => t.status === models.TestResultStatus.ERROR || t.status === models.TestResultStatus.SKIPPED
+    );
   }
 }